+++ backend/src/models/__init__.py
@@ -31,6 +31,8 @@ from .organization import (  # noqa: F401
     Position,
 )
 from .ownership import Ownership  # noqa: F401
+from .project import Project  # noqa: F401
+from .project_relations import ProjectOwnershipRelation  # noqa: F401
 from .property_certificate import (  # noqa: F401
     CertificateType,
     OwnerType,